backend/requirements.txt
flask
flask-cors
requests
python-dotenv
pydub
numpy
soundfile
webrtcvad
SpeechRecognition
google-generativeai

